Renamed file names for merge conflict
